Fix plural

+++ packages/twenty-front/src/modules/settings/admin-panel/health-status/graphql/queries/getQueueJobs.ts
@@ -4,8 +4,8 @@ export const GET_QUEUE_JOBS = gql`
   query GetQueueJobs(
     $queueName: String!
     $state: JobState
-    $limit: Float
-    $offset: Float
+    $limit: Int
+    $offset: Int
   ) {
     getQueueJobs(
       queueName: $queueName
